# Exploratory concept for ELRS interrupt/busy handling