--- a/src/main/drivers/bus_spi_stdperiph.c
+++ b/src/main/drivers/bus_spi_stdperiph.c
@@ -215,9 +215,6 @@ void spiInternalInitStream(const extDevice_t *dev, bool preInit)
 
 void spiInternalStartDMA(const extDevice_t *dev)
 {
-    // Assert Chip Select
-    IOLo(dev->busType_u.spi.csnPin);
-
     dmaChannelDescriptor_t *dmaTx = dev->bus->dmaTx;
     dmaChannelDescriptor_t *dmaRx = dev->bus->dmaRx;
     DMA_Stream_TypeDef *streamRegsTx = (DMA_Stream_TypeDef *)dmaTx->ref;
@@ -355,7 +352,7 @@ void spiSequenceStart(const extDevice_t *dev)
     SPI_Cmd(instance, ENABLE);
 
     // Check that any there are no attempts to DMA to/from CCD SRAM
-    for (busSegment_t *checkSegment = bus->curSegment; checkSegment->len; checkSegment++) {
+    for (busSegment_t *checkSegment = (busSegment_t *)bus->curSegment; checkSegment->len; checkSegment++) {
         // Check there is no receive data as only transmit DMA is available
         if (((checkSegment->u.buffers.rxData) && (IS_CCM(checkSegment->u.buffers.rxData) || (bus->dmaRx == (dmaChannelDescriptor_t *)NULL))) ||
             ((checkSegment->u.buffers.txData) && IS_CCM(checkSegment->u.buffers.txData))) {
@@ -367,17 +364,25 @@ void spiSequenceStart(const extDevice_t *dev)
         xferLen += checkSegment->len;
     }
     // Use DMA if possible
-    if (bus->useDMA && dmaSafe && ((segmentCount > 1) || (xferLen >= 8))) {
+    // If there are more than one segments, or a single segment with negateCS negated then force DMA irrespective of length
+    if (bus->useDMA && dmaSafe && ((segmentCount > 1) || (xferLen >= 8) || !bus->curSegment->negateCS)) {
         // Intialise the init structures for the first transfer
         spiInternalInitStream(dev, false);
+
+        // Assert Chip Select
+        IOLo(dev->busType_u.spi.csnPin);
 
         // Start the transfers
         spiInternalStartDMA(dev);
     } else {
+        busSegment_t *lastSegment = NULL;
+
         // Manually work through the segment list performing a transfer for each
         while (bus->curSegment->len) {
-            // Assert Chip Select
-            IOLo(dev->busType_u.spi.csnPin);
+            if (!lastSegment || lastSegment->negateCS) {
+                // Assert Chip Select if necessary - it's costly so only do so if necessary
+                IOLo(dev->busType_u.spi.csnPin);
+            }
 
             spiInternalReadWriteBufPolled(
                     bus->busType_u.spi.instance,
@@ -407,14 +412,20 @@ void spiSequenceStart(const extDevice_t *dev)
                     break;
                 }
             }
+            lastSegment = (busSegment_t *)bus->curSegment;
             bus->curSegment++;
+        }
+
+        if (lastSegment && !lastSegment->negateCS) {
+            // Negate Chip Select if not done so already
+            IOHi(dev->busType_u.spi.csnPin);
         }
 
         // If a following transaction has been linked, start it
         if (bus->curSegment->u.link.dev) {
             const extDevice_t *nextDev = bus->curSegment->u.link.dev;
-            busSegment_t *nextSegments = bus->curSegment->u.link.segments;
-            busSegment_t *endSegment = bus->curSegment;
+            busSegment_t *nextSegments = (busSegment_t *)bus->curSegment->u.link.segments;
+            busSegment_t *endSegment = (busSegment_t *)bus->curSegment;
             bus->curSegment = nextSegments;
             endSegment->u.link.dev = NULL;
             spiSequenceStart(nextDev);