--- a/src/main/drivers/rx/rx_spi.h
+++ b/src/main/drivers/rx/rx_spi.h
@@ -24,12 +24,14 @@
 
 #include "common/time.h"
 
+#include "drivers/bus.h"
 #include "drivers/exti.h"
 
 #define RX_SPI_MAX_PAYLOAD_SIZE 35
 
 struct rxSpiConfig_s;
 
+extDevice_t *rxSpiGetDevice(void);
 void rxSpiDevicePreInit(const struct rxSpiConfig_s *rxSpiConfig);
 bool rxSpiDeviceInit(const struct rxSpiConfig_s *rxSpiConfig);
 void rxSpiSetNormalSpeedMhz(uint32_t mhz);
@@ -43,6 +45,7 @@ void rxSpiWriteCommandMulti(uint8_t command, const uint8_t *data, uint8_t length
 uint8_t rxSpiReadCommand(uint8_t command, uint8_t commandData);
 void rxSpiReadCommandMulti(uint8_t command, uint8_t commandData, uint8_t *retData, uint8_t length);
 void rxSpiExtiInit(ioConfig_t rxSpiExtiPinConfig, extiTrigger_t rxSpiExtiPinTrigger);
+void rxSpiEnableExti(void);
 bool rxSpiExtiConfigured(void);
 bool rxSpiGetExtiState(void);
 bool rxSpiPollExti(void);