… src/main/drivers/rx/rx_sx1280.c
… src/main/drivers/rx/rx_sx1280.h
… src/main/fc/core.c
… src/main/fc/core.h
… src/main/fc/rc_controls.c
… src/main/fc/rc_controls.h
… src/main/fc/tasks.c
… src/main/flight/failsafe.c
… src/main/flight/failsafe.h
… src/main/rx/expresslrs.c
… src/main/rx/expresslrs.h
… src/main/rx/expresslrs_common.c
… src/main/rx/expresslrs_common.h
… src/main/rx/expresslrs_impl.h
… src/main/rx/rx.c
… src/main/rx/rx.h
… src/main/rx/rx_spi.c
… src/main/scheduler/scheduler.c
… src/main/startup/startup_stm32g474xx.s
… src/main/target/CRAZYBEEF4SX1280/target.h
… src/test/unit/link_quality_unittest.cc
… src/test/unit/rx_ranges_unittest.cc
… src/test/unit/rx_rx_unittest.cc
… src/test/unit/rx_spi_expresslrs_unittest.cc
… src/test/unit/scheduler_unittest.cc
… unified_targets/configs/CRAZYBEEF4SX1280.config