@SteveCEvans 2022-02-03
Latest push demonstrates the following executed all under interrupt control.
1. A read of the SX1280's `SX1280_RADIO_GET_IRQSTATUS` register
2. A write to the `SX1280_RADIO_CLR_IRQSTATUS` to clear the interrupt
3. Waiting on an interrupt to indicate that the busy line has been negated following this

This now uses 4.6us of task time instead of the 28.7us that it would have taken using the old polled approach.

![image](https://user-images.githubusercontent.com/11480839/152437526-c265ee3f-fa5d-4ca0-b872-2113fbe122f9.png)