@phobos- 2022-02-04
Thanks for the huge effort @SteveCEvans I'll be testing this over the weekend. It for sure makes more sense than timeout implementation, however I went with this approach as it is also done the same way in the original code base and I wanted to be sure it's working before changing the approach.