@SteveCEvans 2022-02-10
This code is now working and worthy of a full review. Please also bench test if possible.

This development was done on a BETAFPVF4SX1280 and the traces below are from that FC running a 4K PID loop.

On master we get the following timings.

- `frame reception` shows the time taken from SX1280 interrupt to it being acknowledged.
- `rxDataProcessingRequired()/taskUpdateRxMain()` shows (first pulse) the call to `rxDataProcessingRequired()` and subsequent calls to `taskUpdateRxMain()`.
- `busy` shows the time spent busy polling the SX1280.
- `schedulerExecuteTask` shows when tasks are being executed.

1. Timing markers `P0` to `P4` show the disturbance in the gyro loop, resulting in a long 234.5us period followed by short 44.3us and 96.6us periods before recovering.
2. `P5` shows a latency of 964us from the SX1280 interrupt to the `RX` task processing completing.
3. `P6` shows time taken from the SX1280 interrupt to the call to `rxDataProcessingRequired()` of 219us.
4. `P7` shows time taken from the SX1280 interrupt to the last busy pull to be 208us.

![image](https://user-images.githubusercontent.com/11480839/153490077-6b8150f3-e2f3-47a7-9cd4-394e6fc9b25b.png)

With this PR we get the following timings.

- `frame reception` shows the time taken from SX1280 interrupt to it being acknowledged.
- `rxDataProcessingRequired()/taskUpdateRxMain()` shows (first pulse) the call to `rxDataProcessingRequired()` and subsequent calls to `taskUpdateRxMain()`.
- `busy` shows the time spent waiting for the SX1280 whilst busy, but this is now done using an interrupt so the processor can execute other tasks whilst waiting.
- `schedulerExecuteTask` shows when tasks are being executed. Note that unlike master there are two tasks running in the background whilst the SX1280 is accessed.

1. Timing markers `P0` to `P4` show the disturbance in the gyro loop, resulting in a long 131.5us period followed by short 123.6 us and 118.4us periods before recovering. This is clearly less disturbance than on master.
2. `P5` shows a latency of 634us from the SX1280 interrupt to the `RX` task processing completing. This PR removes a state from the RX task as some processing has been moved out of the checker to `rxFrameCheck()`.
3. `P6` shows time taken from the SX1280 interrupt to the call to `rxDataProcessingRequired()` of 151us.
4. `P7` shows time taken from the SX1280 interrupt to the last busy pull to be 108us.

![image](https://user-images.githubusercontent.com/11480839/153491699-32263844-681d-4138-941f-e07584d9b921.png)

As the SX1280 processing runs in the background under interrupt control there are no longer any disconnects when running commands in the CLI.

Saving to FLASH, either from the configurator or as a result of using stick commands stalls all interrupt activity and thus still results in the connection being dropped. I'm looking into that.