@haslinghuis 2022-02-11
For testing:
[betaflight_4.3.0_BETAFPVF4SX1280_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045129/betaflight_4.3.0_BETAFPVF4SX1280_afe52466d.zip)
[betaflight_4.3.0_CRAZYBEEF4SX1280_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045130/betaflight_4.3.0_CRAZYBEEF4SX1280_afe52466d.zip)
[betaflight_4.3.0_STM32H743_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045131/betaflight_4.3.0_STM32H743_afe52466d.zip)
[betaflight_4.3.0_STM32G47X_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045132/betaflight_4.3.0_STM32G47X_afe52466d.zip)
[betaflight_4.3.0_STM32F745_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045133/betaflight_4.3.0_STM32F745_afe52466d.zip)
[betaflight_4.3.0_STM32F7X2_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045134/betaflight_4.3.0_STM32F7X2_afe52466d.zip)
[betaflight_4.3.0_STM32F411_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045135/betaflight_4.3.0_STM32F411_afe52466d.zip)
[betaflight_4.3.0_STM32F405_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045136/betaflight_4.3.0_STM32F405_afe52466d.zip)