@mard89 2022-02-12
> For testing: [betaflight_4.3.0_BETAFPVF4SX1280_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045129/betaflight_4.3.0_BETAFPVF4SX1280_afe52466d.zip) [betaflight_4.3.0_CRAZYBEEF4SX1280_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045130/betaflight_4.3.0_CRAZYBEEF4SX1280_afe52466d.zip) [betaflight_4.3.0_STM32H743_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045131/betaflight_4.3.0_STM32H743_afe52466d.zip) [betaflight_4.3.0_STM32G47X_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045132/betaflight_4.3.0_STM32G47X_afe52466d.zip) [betaflight_4.3.0_STM32F745_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045133/betaflight_4.3.0_STM32F745_afe52466d.zip) [betaflight_4.3.0_STM32F7X2_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045134/betaflight_4.3.0_STM32F7X2_afe52466d.zip) [betaflight_4.3.0_STM32F411_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045135/betaflight_4.3.0_STM32F411_afe52466d.zip) [betaflight_4.3.0_STM32F405_afe52466d.zip](https://github.com/betaflight/betaflight/files/8045136/betaflight_4.3.0_STM32F405_afe52466d.zip)

This build greatly aggravated the problem I was seeing on my 12a BETAFPV4SX1280 board. Where as the failsafes were previously spaced about 10 to 15 minutes apart, I now get multiple failsafes per minute.

![Screenshot_20220211_091046](https://user-images.githubusercontent.com/51527720/153703708-fd5f8f0a-4d5e-4925-9f85-59b869209b0e.png)