@CMartin249 2022-02-13
Tested betaflight_4.3.0_CRAZYBEEF4SX1280_afe52466d.hex and had to go back to RC3. Lots of RXLOST. Couldn't get 15 feet. Connection also couldn't reconnect on its own most of time; had to power cycle. Running 8K/2K and without bidir dshot enabled.