@ctzsnooze 2022-02-15
Same here, repetitively lost signal.