@SteveCEvans 2022-02-20
@mard89 @CMartin249 @ctzsnooze @AlessandroAU @phobos- Please test this latest push. I believe it to be functional. Images below. I'm now looking at how to make the link survive EEPROM writes.

[betaflight_4.3.0_BETAFPVF4SX1280_aa3cb87.hex.zip](https://github.com/betaflight/betaflight/files/8104987/betaflight_4.3.0_BETAFPVF4SX1280_aa3cb87.hex.zip)
[betaflight_4.3.0_CRAZYBEEF4SX1280_aa3cb87.hex.zip](https://github.com/betaflight/betaflight/files/8104991/betaflight_4.3.0_CRAZYBEEF4SX1280_aa3cb87.hex.zip)